backend/requirements.txt
django
djangorestframework
PyJWT == 2.1.0
djangorestframework-simplejwt == 4.8.0
pillow
drf-yasg
django-allauth
dj_rest_auth
django-cors-headers
gevent
boto3

celery
redis
django_celery_beat
django_celery_results

coverage

hypothesis
hypothesis[all]

gunicorn

psycopg2-binary

celery

#Yolo v5 requirements.txt
# base ----------------------------------------
# matplotlib>=3.2.2
# numpy>=1.18.5
opencv-python>=4.1.2
# Pillow
PyYAML==5.4.1
# scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0
cv

# plotting ------------------------------------
seaborn>=0.11.0
pandas
djongo
